Add basic skeleton of profile editor dialog
