+++ lib/widgets/widgets.dart
@@ -2,5 +2,5 @@ export './sign_in.dart';
 export './register.dart';
 export './app_wrapper.dart';
 export './expandable_datatable.dart';
-export './profile_dialog.dart';
+export 'profile_editor.dart';
 export './raid_form.dart';
